rydde kode Co-authored-by: sasoria <sergio.alberto.arevalo.soria@nav.no>
