+++ src/hooks/useOversikt.ts
@@ -26,3 +26,5 @@ export const useOversikt = (produktProperties?: ProduktProperties[]) => {
   );
 };
 
+
+
